+++ liblzma-rs/src/common/index_hash.rs
@@ -328,7 +328,7 @@ pub unsafe extern "C" fn lzma_index_hash_decode(
             }
             let fresh2 = *in_pos;
             *in_pos = (*in_pos).wrapping_add(1);
-            if (*index_hash).crc32 >> (*index_hash).pos.wrapping_mul(8) & 0xff as u32
+            if (*index_hash).crc32 >> (*index_hash).pos.wrapping_mul(8) & 0xff
                 != *in_0.offset(fresh2 as isize) as u32
             {
                 return LZMA_DATA_ERROR;
